[addonsinfoprovider] rename addon
