+++ bundles/org.openhab.misc.addonsinfoprovider/.gitignore
