+++ bundles/org.openhab.misc.addonsinfoprovider/NOTICE
