+++ bundles/org.openhab.misc.addonsinfoprovider/README.md
@@ -1,4 +1,4 @@
-# AddonSuggestionFinder Addon
+# AddonsInfoProvider Addon
 
 This is a special addon that implements an `AddonInfoProvider` service containing information about suggested Addons that could potentially be installed.
 It allows developers to include information in their own addons so that the system can scan the user's network to discover potential addons that can automatically be installed.
