+++ bundles/org.openhab.misc.addonsinfoprovider/src/main/resources/OH-INF/addon/addon.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<addon:addon id="addonsinfoprovider" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns:addon="https://openhab.org/schemas/addon/v1.0.0"
+	xsi:schemaLocation="https://openhab.org/schemas/addon/v1.0.0 https://openhab.org/schemas/addon-1.0.0.xsd">
+
+	<type>misc</type>
+	<name>Addons Information Provider</name>
+	<description>This is a special addon for providing information about other addons to be installed on setup.</description>
+
+</addon:addon>
