+++ bundles/org.openhab.misc.addonsinfoprovider/src/main/resources/footer.xml
