+++ bundles/org.openhab.misc.addonsinfoprovider/src/main/resources/header.xml
