--- bundles/org.openhab.misc.addonsuggestionfinder/src/main/resources/OH-INF/addon/addon.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<addon:addon id="addonsuggestionfinder" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:addon="https://openhab.org/schemas/addon/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/addon/v1.0.0 https://openhab.org/schemas/addon-1.0.0.xsd">
-
-	<type>misc</type>
-	<name>Addon Suggestion Finder</name>
-	<description>This is a special addon for finding suggested addons to be installed on setup.</description>
-
-</addon:addon>
